+++ OnAct/Controllers/GroupsController.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using OnAct.Auth.Model;
 using OnAct.Data.Dtos.Groups;
 using OnAct.Data.Dtos.Places;
 using OnAct.Data.Entities;
@@ -15,15 +17,17 @@ namespace OnAct.Controllers
         private readonly IMapper _mapper;
         private readonly IActivitiesRepository _activitiesRepository;
         private readonly IPlacesRepository _placesRepository;
+        private readonly IAuthorizationService _authorizationService;
 
 
         public GroupsController(IGroupsRepository groupsRepository, IMapper mapper,
-            IActivitiesRepository activitiesRepository, IPlacesRepository placesRepository)
+            IActivitiesRepository activitiesRepository, IPlacesRepository placesRepository, IAuthorizationService authorizationService)
         {
             _groupsRepository = groupsRepository;
             _mapper = mapper;
             _activitiesRepository = activitiesRepository;
             _placesRepository = placesRepository;
+            _authorizationService = authorizationService;
         }
 
 
@@ -51,6 +55,7 @@ namespace OnAct.Controllers
         }
 
         [HttpPost]
+        [Authorize(Policy = "CreatorRights")]
         public async Task<ActionResult<GroupDto>> PostAsync(int activityId, int placeId, CreateGroupDto groupDto)
         {
             var activity = await _activitiesRepository.Get(activityId);
@@ -59,10 +64,19 @@ namespace OnAct.Controllers
             var place = await _placesRepository.Get(activityId, placeId);
             if (place == null) return NotFound($"Place with id '{placeId}' not found in activity with id '{activityId}'.");
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, activity, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                //404
+                return Forbid();
+            }
+
             var group = _mapper.Map<Group>(groupDto);
             group.ActivityId = activityId;
             group.PlaceId = placeId;
             group.CreatedDate = DateTime.UtcNow;
+
 
             await _groupsRepository.Create(group);
 
@@ -70,6 +84,7 @@ namespace OnAct.Controllers
         }
 
         [HttpPut("{groupId}")]
+        [Authorize(Policy = "CreatorRights")]
         public async Task<ActionResult<GroupDto>> PutAsync(int activityId, int placeId, int groupId, UpdateGroupDto groupDto)
         {
             var activity = await _activitiesRepository.Get(activityId);
@@ -81,6 +96,14 @@ namespace OnAct.Controllers
             var oldGroup = await _groupsRepository.Get(activityId, placeId, groupId);
             if (oldGroup == null) return NotFound($"Group with id '{groupId}' not found in activity with id '{activityId}' and place with id '{placeId}'.");
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, activity, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                //404
+                return Forbid();
+            }
+
             //oldPost.Body = postDto.Body;
             _mapper.Map(groupDto, oldGroup);
 
@@ -90,6 +113,7 @@ namespace OnAct.Controllers
         }
 
         [HttpDelete("{groupId}")]
+        [Authorize(Policy = "CreatorRights")]
         public async Task<ActionResult> DeleteAsync(int activityId, int placeId, int groupId)
         {
             var activity = await _activitiesRepository.Get(activityId);
@@ -101,6 +125,13 @@ namespace OnAct.Controllers
             var group = await _groupsRepository.Get(activityId, placeId, groupId);
             if (group == null) return NotFound($"Group with id '{groupId}' not found in activity with id '{activityId}' and place with id '{placeId}'.");
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, activity, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                //404
+                return Forbid();
+            }
 
             await _groupsRepository.Delete(group);
